# Conflicts:
#	src/main/java/com/ssibssaggi/findex/domain/entity/index/IndexInformation.java
#	src/main/java/com/ssibssaggi/findex/repository/IndexInformationRepository.java

+++ src/main/java/com/ssibssaggi/findex/controller/dto/IndexInformationUpdateRequest.java
@@ -3,7 +3,7 @@ package com.ssibssaggi.findex.controller.dto;
 public record IndexInformationUpdateRequest(
         Integer employedItemsCount,
         String basePointInTime,
-        Integer baseIndex,
+        Float baseIndex,
         Boolean favorite
 ) {
 }
